+++ rs/src/morph_shape.rs
@@ -135,6 +135,7 @@ pub(crate) fn emit_morph_style_change_bits<W: WriteBits>(
   let has_new_line_style = value.line_style.is_some();
   let has_new_styles = value.new_styles.is_some();
 
+  #[allow(clippy::identity_op)]
   let flags: u8 = 0
     | (if has_move_to { 1 << 0 } else { 0 })
     | (if has_new_left_fill { 1 << 1 } else { 0 })
@@ -187,8 +188,11 @@ pub(crate) fn emit_morph_fill_style<W: io::Write + ?Sized>(
 ) -> io::Result<()> {
   match value {
     ast::MorphFillStyle::Bitmap(ref style) => {
-      let code: u8 =
-        0 | (if !style.repeating { 1 << 0 } else { 0 }) | (if !style.smoothed { 1 << 1 } else { 0 }) | 0x40;
+      #[allow(clippy::identity_op)]
+      let code: u8 = 0
+        | (if !style.repeating { 1 << 0 } else { 0 })
+        | (if !style.smoothed { 1 << 1 } else { 0 })
+        | 0x40;
       emit_u8(writer, code)?;
       emit_morph_bitmap_fill(writer, style)
     }
@@ -294,14 +298,12 @@ pub(crate) fn emit_morph_line_style2<W: io::Write + ?Sized>(
   emit_le_u16(writer, value.width)?;
   emit_le_u16(writer, value.morph_width)?;
 
-  let has_fill = match &value.fill {
-    ast::MorphFillStyle::Solid(_) => false,
-    _ => true,
-  };
+  let has_fill = !matches!(&value.fill, ast::MorphFillStyle::Solid(_));
   let join_style_code = join_style_to_code(value.join);
   let start_cap_style_code = cap_style_to_code(value.start_cap);
   let end_cap_style_code = cap_style_to_code(value.end_cap);
 
+  #[allow(clippy::identity_op)]
   let flags: u16 = 0
     | (if value.pixel_hinting { 1 << 0 } else { 0 })
     | (if value.no_v_scale { 1 << 1 } else { 0 })